requirements.txt
requests
pandas
finnhub-python
python-dotenv
pyTelegramBotAPI